# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api integração connector de emails outlook",
        "title": "outlook-connector",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/v1/connector/read": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Obter os emails para os parametros informados.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "Leitura de Emails",
                "operationId": "GetEmailFilter",
                "parameters": [
                    {
                        "description": "Requisição Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetEmailFilterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados recebidos!",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            }
        },
        "/v1/open/token": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Obter os dados de token JWT para acesso.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get Token",
                "operationId": "GetToken",
                "parameters": [
                    {
                        "description": "Requisição Body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados recebidos!",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    },
                    "400": {
                        "description": "Requisição Inválida",
                        "schema": {
                            "$ref": "#/definitions/response.HttpResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "request.GetEmailFilterRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Flag para retornar a contagem total",
                    "type": "boolean"
                },
                "expand": {
                    "description": "Propriedades adicionais para expandir",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "filter": {
                    "description": "Condição de filtro",
                    "type": "string"
                },
                "includeHiddenMessages": {
                    "description": "Flag para incluir mensagens ocultas",
                    "type": "boolean"
                },
                "orderby": {
                    "description": "Ordenação dos resultados",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "search": {
                    "description": "Termo de pesquisa",
                    "type": "string"
                },
                "select": {
                    "description": "Seleção de campos específicos",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "skip": {
                    "description": "Número de registros para ignorar",
                    "type": "integer"
                },
                "top": {
                    "description": "Número máximo de registros a retornar",
                    "type": "integer"
                }
            }
        },
        "request.GetTokenRequest": {
            "type": "object",
            "required": [
                "key"
            ],
            "properties": {
                "key": {
                    "type": "string"
                }
            }
        },
        "response.HttpResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}